feat(fish): update 'upgrade_libraries' for multiple package managers

+++ fish/.config/fish/functions/upgrade_libraries.fish
@@ -1,6 +1,16 @@
 function upgrade_libraries
-    brew update
-    brew upgrade
-    brew autoremove
-    brewdump
+    if command -q brew
+        brew update
+        brew upgrade
+        brew autoremove
+        brewdump
+    else if command -q pacman
+        sudo pacman -Syu
+    else if command -q apt
+        sudo apt update
+        sudo apt upgrade -y
+    else
+        echo 'Define how to upgrade libraries'
+        exit 1
+    end
 end
